Merge for service version 2.0 release

+++ app/scripts/app.js
@@ -11,6 +11,7 @@ var VECTOR_BREAKDOWN = {};
     root.define([
             'backbone',
             'globals',
+            'cesium',
             'regions/DialogRegion', 'regions/UIRegion',
             'layouts/LayerControlLayout',
             'layouts/ToolControlLayout',
@@ -28,7 +29,7 @@ var VECTOR_BREAKDOWN = {};
             'controller/DataController'
         ],
 
-        function(Backbone, globals, DialogRegion,
+        function(Backbone, globals, Cesium, DialogRegion,
             UIRegion, LayerControlLayout, ToolControlLayout, OptionsLayout,
             WindowView, Communicator, FilePond) {
 
@@ -129,26 +130,69 @@ var VECTOR_BREAKDOWN = {};
 
                 // Check if version of service is set and if it differs from the
                 // current version
+                var serviceVersion;
                 if(localStorage.getItem('serviceVersion') !== null){
-                    var serviceVersion = JSON.parse(
+                    serviceVersion = JSON.parse(
                         localStorage.getItem('serviceVersion')
                     );
                     if(serviceVersion!==globals.version){
-                        // A new version has been loaded, here we could 
-                        // differentiate which version was previous and which
-                        // one ist the new, for now we reset and save the new
-                        // version
-                        showMessage('success',
-                            'A new version ('+globals.version+') of the service has been released. '+
-                            'Your configuration has been updated.</br>'+
-                            'You can find information on the changes in the '+
-                            '<b><a target="_blank" href="/accounts/changelog">changelog</a></b>.', 35
-                        );
-                        localStorage.clear();
-                        localStorage.setItem(
-                            'serviceVersion',
-                            JSON.stringify(globals.version)
-                        );
+                        if(localStorage.getItem('configurationLoaded') !== null){
+                            // A configuration from a previous version was loaded
+                            localStorage.removeItem('configurationLoaded');
+                            // If service version older then 1.3 it should not be
+                            // loaded anyhow, but we add check here just in case
+                            var numberSV = Number(serviceVersion);
+                            if(isNaN(numberSV) || numberSV<1.3){
+                                localStorage.clear();
+                                showMessage('success',
+                                    'The configuration you are trying to load is outdated.', 35
+                                );
+                            } else if(numberSV === 1.4){
+                                // In version 1.4 all parameters are requested
+                                // so we need to identify which parameter
+                                // are used to make sure they are available
+                                var usedPars = [
+                                    JSON.parse(localStorage.getItem('xAxisSelection')),
+                                    JSON.parse(localStorage.getItem('yAxisSelection')),
+                                    JSON.parse(localStorage.getItem('y2AxisSelection')),
+                                    JSON.parse(localStorage.getItem('colorAxisSelection')),
+                                    JSON.parse(localStorage.getItem('colorAxis2Selection'))
+                                ].flat(2);
+
+                                var dataSettings = JSON.parse(localStorage.getItem('dataSettings'));
+                                for (var uu = 0; uu < usedPars.length; uu++) {
+                                    if(usedPars[uu] !== null){
+                                        // Look for parameter in product config
+                                        if(dataSettings.hasOwnProperty(usedPars[uu])){
+                                            dataSettings[usedPars[uu]].active = true;
+                                        }
+                                    }
+                                }
+                                localStorage.setItem('dataSettings', JSON.stringify(dataSettings));
+                            } else {
+                                localStorage.setItem(
+                                    'serviceVersion',
+                                    JSON.stringify(globals.version)
+                                );
+                            }
+
+                        } else {
+                            // A new version has been loaded, here we could 
+                            // differentiate which version was previous and which
+                            // one ist the new, for now we reset and save the new
+                            // version
+                            showMessage('success',
+                                'A new version ('+globals.version+') of the service has been released. '+
+                                'Your configuration has been updated.</br>'+
+                                'You can find information on the changes in the '+
+                                '<b><a target="_blank" href="/accounts/changelog">changelog</a></b>.', 35
+                            );
+                            localStorage.clear();
+                            localStorage.setItem(
+                                'serviceVersion',
+                                JSON.stringify(globals.version)
+                            );
+                        }
                     }
                 } else {
                     // This should be the case when loading version 2.3 for the 
@@ -265,7 +309,7 @@ var VECTOR_BREAKDOWN = {};
                         }
 
                         // Check if there are changes saved for parameter settings
-                        if(localStorage.getItem('dataSettings') !== null){
+                        if(localStorage.getItem('dataSettings') !== 'null'){
                             var setts = JSON.parse(localStorage.getItem('dataSettings'));
                             for (var key in product_config[i].parameters){
                                 if(setts.hasOwnProperty(key)){
@@ -386,6 +430,94 @@ var VECTOR_BREAKDOWN = {};
                     
                     console.log("Added product " + product.name );
                 }, this);
+
+                // Check if datasettings already available, and also from which
+                // service version the datasettings are coming from
+                var numberSV = Number(serviceVersion);
+
+                if(!isNaN(numberSV) && numberSV>1.4 && 
+                    localStorage.getItem('dataSettings') !== null){
+                    globals.dataSettings = JSON.parse(localStorage.getItem('dataSettings'));
+                } else {
+                    // Go through products and fill global datasettings
+                    globals.products.each(function(product){
+                        var downloadPars = product.get('download_parameters');
+                        var productConf = product.get('parameters');
+                        var prodId = product.get('download').id;
+                        var prevDataSettings = JSON.parse(localStorage.getItem('dataSettings'));
+                        // Get keys we currently request for visualization
+                        var downloadKeys = [];
+
+                        if(globals.fieldList.hasOwnProperty(prodId)){
+                            if(Array.isArray(globals.fieldList[prodId])){
+                                downloadKeys = globals.fieldList[prodId];
+                            } else {
+                                for (var key in globals.fieldList[prodId]){
+                                    downloadKeys = downloadKeys.concat(
+                                        globals.fieldList[prodId][key]
+                                    );
+                                }
+                            }
+                        }
+                        var isAux = prodId.indexOf('AUX_')!==-1;
+                        for(var key in downloadPars){
+                            if(!globals.dataSettings[prodId].hasOwnProperty(key) && 
+                                downloadKeys.indexOf(key)!==-1){
+                                var parInfo = {};
+                                // Check if parameter defined in product config
+                                if(productConf.hasOwnProperty(key)){
+                                    for(var prk in productConf[key]){
+                                        parInfo[prk] = productConf[key][prk];
+                                    }
+                                }
+                                if(downloadPars[key].hasOwnProperty('uom')){
+                                    parInfo.uom = downloadPars[key].uom;
+                                }
+                                if(downloadPars[key].hasOwnProperty('required')){
+                                    parInfo.active = true;
+                                }
+                                if(downloadPars[key].hasOwnProperty('active')){
+                                    parInfo.active = downloadPars[key].active;
+                                }
+                                // For now we activate all auxiliary parameters per 
+                                // default
+                                // TODO: Will need to change once 2D parameters are introduced
+                                if(isAux && prodId!=='AUX_MET_12'){
+                                    parInfo.active = true;
+                                }
+                                globals.dataSettings[prodId][key] = parInfo;
+                            } else if(downloadKeys.indexOf(key)!==-1){
+                                // Add info if available in product config
+                                if(productConf.hasOwnProperty(key)){
+                                    for(var prk in productConf[key]){
+                                        if(productConf[key][prk]!==null){
+                                            globals.dataSettings[prodId][key][prk] = productConf[key][prk];
+                                        }
+                                    }
+                                }
+                                // Add active info if already present
+                                if(downloadPars[key].hasOwnProperty('required')){
+                                    globals.dataSettings[prodId][key].active = true;
+                                }
+                                if(downloadPars[key].hasOwnProperty('active')){
+                                    globals.dataSettings[prodId][key].active = downloadPars[key].active;
+                                }
+                                // For now we activate all auxiliary parameters per 
+                                // default
+                                if(isAux && prodId!=='AUX_MET_12'){
+                                    globals.dataSettings[prodId][key].active = true;
+                                }
+                            }
+                            if(prevDataSettings && prevDataSettings.hasOwnProperty(key)){
+                                if(prevDataSettings[key].hasOwnProperty('active')){
+                                    globals.dataSettings[prodId][key].active = true;
+                                }
+                            }
+                        }
+                    });
+                    localStorage.setItem('dataSettings', JSON.stringify(globals.dataSettings));
+                    localStorage.setItem('serviceVersion', JSON.stringify(globals.version));
+                }
 
                 var productcolors = d3.scale.ordinal().domain(domain).range(range);
 
@@ -574,6 +706,9 @@ var VECTOR_BREAKDOWN = {};
                     })
                 });
 
+                // Create data configuration view
+                this.dataConfigurationView = new v.DataConfigurationView();
+
 
                 this.layerSettings = new v.LayerSettings();
 
@@ -719,8 +854,6 @@ var VECTOR_BREAKDOWN = {};
 
                 // Remove loading screen when this point is reached in the script
                 $('#loadscreen').remove();
-
-
             }
